--- a/eval/memo.go
+++ b/eval/memo.go
@@ -4,7 +4,7 @@ import (
 	"grol.io/grol/object"
 )
 
-const MaxArgs = 6
+const MaxArgs = 4
 
 type CacheKey struct {
 	Fn   string
@@ -18,6 +18,9 @@ func NewCache() Cache {
 }
 
 func (c Cache) Get(fn string, args []object.Object) (object.Object, bool) {
+	if len(args) > MaxArgs {
+		return nil, false
+	}
 	key := CacheKey{Fn: fn}
 	for i, v := range args {
 		// Can't hash functions arguments (yet).
@@ -31,6 +34,9 @@ func (c Cache) Get(fn string, args []object.Object) (object.Object, bool) {
 }
 
 func (c Cache) Set(fn string, args []object.Object, result object.Object) {
+	if len(args) > MaxArgs {
+		return
+	}
 	key := CacheKey{Fn: fn}
 	for i, v := range args {
 		// Can't hash functions arguments (yet).