--- a/examples/fib.gr
+++ b/examples/fib.gr
@@ -1,0 +1,10 @@
+fib = func(x) {
+	if x <= 0 {
+		return 0
+	}
+	if x == 1 {
+		return 1
+	}
+	fib(x-1) + fib(x-2)
+}
+fib(50)