--- a/examples/fib.gr
+++ b/examples/fib.gr
@@ -5,8 +5,8 @@ fib = func(x) {
 	if x == 1 {
 		return 1
 	}
-	fib(x-1) + fib(x-2)
+	fib(x - 1) + fib(x - 2)
 }
-r = fib(50)
-log("fib(50) =", r)
+r = fib(35)
+log("fib(35) =", r)
 r