Fixes #72

before
```
BenchmarkGrol/Fibonacci-10.gr-11         	    7478	    155847 ns/op	  130851 B/op	    6502 allocs/op
BenchmarkGrol/Fibonacci-35.gr-11         	       1	26592782750 ns/op	22056347384 B/op	1096181911 allocs/op
```

after
```
BenchmarkGrol/Fibonacci-10.gr-11         	   83880	     14579 ns/op	   13966 B/op	     559 allocs/op
BenchmarkGrol/Fibonacci-35.gr-11         	   24448	     48817 ns/op	   52861 B/op	    1834 allocs/op
```

- [x] handle `print`

We handle print by capturing and saving the output - we exclude log() even in "EvalString" (wasm/discord) mode
so log will show real execution while print will work like expected if there was no memoization

- [x] semi unrelated but added Share button, best used after checking "compact" and run, then share for a short url with the function